fix: delete files that are no longer needed

+++ packages/backend/src/queue/processors/ImportCustomAvatarDecorationsProcessorService.ts
@@ -17,6 +17,8 @@ import { bindThis } from '@/decorators.js';
 import { QueueLoggerService } from '../QueueLoggerService.js';
 import type * as Bull from 'bullmq';
 import type { DbUserImportJobData } from '../types.js';
+import {ApiError} from "@/server/api/error.js";
+import {meta} from "@/server/api/endpoints/drive/files/delete.js";
 
 // TODO: 名前衝突時の動作を選べるようにする
 @Injectable()
@@ -84,9 +86,20 @@ export class ImportCustomAvatarDecorationsProcessorService {
 					continue;
 				}
 				const adPath = outputPath + '/' + record.fileName;
-				await this.avatarDecorationsRepository.delete({
-					name: adInfo.name,
-				});
+				const adExist = await this.avatarDecorationsRepository.findOneBy({ name: adInfo.name });
+
+				if (adExist) {
+					await this.avatarDecorationsRepository.delete({
+						name: adInfo.name,
+					});
+					const file = await this.driveFilesRepository.findOneBy({ id: adExist.driveId });
+
+					if (file == null) {
+						this.logger.error(`Cannot delete ${ adInfo.id } (${ adInfo.driveId }). Ignored.`);
+						continue;
+					}
+					await this.driveService.deleteFile(file);
+				}
 				try {
 					const driveFile = await this.driveService.addFile({
 						user: null,
@@ -99,6 +112,7 @@ export class ImportCustomAvatarDecorationsProcessorService {
 						url: driveFile.webpublicUrl ?? driveFile.url,
 						description: adInfo.description,
 						roleIdsThatCanBeUsedThisDecoration: [],
+						driveId: driveFile.id,
 					});
 				} catch (e) {
 					if (e instanceof Error || typeof e === 'string') {
